Fix missing fields in schema (#477)
* Fix missing fields in schema

Schema didn’t include the `footerOnly` property needed to filter out Mastodon icon in the header, so it was getting dropped when parsing.

* Infer types from schemas to avoid inconsistencies

Zod ships a `z.infer<>` generic to make it easy to get a type that exactly matches a schema. I’ve used this to replace the manual duplication required to maintain both a TS interface and Zod schema. Should help prevent similar issues in the future where we might update the type but forget to update the schema.