change link
This pull request updates the link for the astro-angular integration on the Astro Integrations page. The previous link pointed to AnalogJS's homepage. It has now been corrected to link directly to the Astro Angular documentation.

The previous link directed users to the AnalogJS homepage, which did not provide specific information about the Astro Angular integration. By updating the link to point directly to the Astro Angular documentation, users can now easily find detailed information and instructions on how to use the integration, improving the overall user experience and accessibility of the documentation.